chore: issues custom context menu
