+++ packages/ui/src/dropdowns/context-menu/root.tsx
@@ -120,7 +120,7 @@ export const ContextMenu: React.FC<ContextMenuProps> = (props) => {
   return (
     <div
       className={cn(
-        "fixed size-full top-0 left-0 cursor-default z-20 opacity-0 pointer-events-none transition-opacity",
+        "fixed h-screen w-screen top-0 left-0 cursor-default z-20 opacity-0 pointer-events-none transition-opacity",
         {
           "opacity-100 pointer-events-auto": isOpen,
         }
